fix moesif data collector
