+++ component/publisher-client/src/main/java/org/wso2/am/analytics/publisher/reporter/GenericInputValidator.java
@@ -67,6 +67,7 @@ import static org.wso2.am.analytics.publisher.util.Constants.USER_NAME;
  * inputs will be validated.
  */
 public class GenericInputValidator {
+
     private static final org.wso2.am.analytics.publisher.reporter.GenericInputValidator INSTANCE =
             new org.wso2.am.analytics.publisher.reporter.GenericInputValidator();
     private static final Map<String, Class> defaultResponseEventSchema = Stream.of(
@@ -86,6 +87,8 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_OWNER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ORGANIZATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ENVIRONMENT_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
@@ -97,38 +100,41 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(BACKEND_LATENCY, Long.class),
                     new AbstractMap.SimpleImmutableEntry<>(REQUEST_MEDIATION_LATENCY, Long.class),
                     new AbstractMap.SimpleImmutableEntry<>(RESPONSE_MEDIATION_LATENCY, Long.class),
-                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class))
+                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
             .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
     private static final Map<String, Class> elkResponseEventSchema = Stream.of(
-            new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(CORRELATION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(KEY_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_METHOD, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_CONTEXT, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_RESOURCE_TEMPLATE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(DESTINATION, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_OWNER, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(USER_NAME, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
-            new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
-            new AbstractMap.SimpleImmutableEntry<>(RESPONSE_CACHE_HIT, Boolean.class),
-            new AbstractMap.SimpleImmutableEntry<>(RESPONSE_LATENCY, Long.class),
-            new AbstractMap.SimpleImmutableEntry<>(BACKEND_LATENCY, Long.class),
-            new AbstractMap.SimpleImmutableEntry<>(REQUEST_MEDIATION_LATENCY, Long.class),
-            new AbstractMap.SimpleImmutableEntry<>(RESPONSE_MEDIATION_LATENCY, Long.class),
-            new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
+                    new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(CORRELATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(KEY_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_METHOD, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_CONTEXT, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_RESOURCE_TEMPLATE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(DESTINATION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_OWNER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ORGANIZATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ENVIRONMENT_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_NAME, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(RESPONSE_CACHE_HIT, Boolean.class),
+                    new AbstractMap.SimpleImmutableEntry<>(RESPONSE_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(BACKEND_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(REQUEST_MEDIATION_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(RESPONSE_MEDIATION_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
             .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
     private static final Map<String, Class> defaultFaultEventSchema = Stream.of(
                     new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
@@ -141,6 +147,35 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(API_TYPE, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_METHOD, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(APPLICATION_OWNER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ORGANIZATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ENVIRONMENT_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(RESPONSE_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
+            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+    private static final Map<String, Class> elkFaultEventSchema = Stream.of(
+                    new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(CORRELATION_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(KEY_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ERROR_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ERROR_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(ERROR_MESSAGE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_ID, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_TYPE, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_METHOD, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
@@ -149,29 +184,10 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
-                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class))
-            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-    private static final Map<String, Class> elkFaultEventSchema = Stream.of(
-            new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(CORRELATION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(KEY_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(ERROR_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(ERROR_CODE, Integer.class),
-            new AbstractMap.SimpleImmutableEntry<>(ERROR_MESSAGE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(APPLICATION_OWNER, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(REGION_ID, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
-            new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
-            new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
-            new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
+                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
             .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
     private static final Map<String, Class> choreoResponseEventSchema = Stream.of(
                     new AbstractMap.SimpleImmutableEntry<>(REQUEST_TIMESTAMP, String.class),
@@ -218,6 +234,7 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(API_NAME, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_VERSION, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_CREATION, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(API_METHOD, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(API_CREATOR_TENANT_DOMAIN, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(APPLICATION_NAME, String.class),
@@ -227,7 +244,11 @@ public class GenericInputValidator {
                     new AbstractMap.SimpleImmutableEntry<>(ENVIRONMENT_ID, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(GATEWAY_TYPE, String.class),
                     new AbstractMap.SimpleImmutableEntry<>(PROXY_RESPONSE_CODE, Integer.class),
-                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class))
+                    new AbstractMap.SimpleImmutableEntry<>(TARGET_RESPONSE_CODE, Integer.class),
+                    new AbstractMap.SimpleImmutableEntry<>(RESPONSE_LATENCY, Long.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_IP, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(USER_AGENT_HEADER, String.class),
+                    new AbstractMap.SimpleImmutableEntry<>(PROPERTIES, LinkedHashMap.class))
             .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
     private static final List<String> configProperties = new ArrayList<>();
 
@@ -236,10 +257,12 @@ public class GenericInputValidator {
     }
 
     public static org.wso2.am.analytics.publisher.reporter.GenericInputValidator getInstance() {
+
         return INSTANCE;
     }
 
     public Map<String, Class> getEventProperties(MetricSchema schema) {
+
         switch (schema) {
             case RESPONSE:
                 return defaultResponseEventSchema;
@@ -259,6 +282,7 @@ public class GenericInputValidator {
     }
 
     public List<String> getConfigProperties() {
+
         return configProperties;
     }
 }
