+++ component/publisher-client/src/main/java/org/wso2/am/analytics/publisher/reporter/moesif/ParallelQueueWorker.java
@@ -19,11 +19,9 @@ package org.wso2.am.analytics.publisher.reporter.moesif;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import org.wso2.am.analytics.publisher.client.MoesifClient;
 import org.wso2.am.analytics.publisher.exception.MetricReportingException;
 import org.wso2.am.analytics.publisher.reporter.MetricEventBuilder;
-import org.wso2.am.analytics.publisher.retriever.MoesifKeyRetriever;
 
 import java.util.concurrent.BlockingQueue;
 
@@ -33,7 +31,6 @@ import java.util.concurrent.BlockingQueue;
 public class ParallelQueueWorker implements Runnable {
     private static final Logger log = LoggerFactory.getLogger(ParallelQueueWorker.class);
     private BlockingQueue<MetricEventBuilder> eventQueue;
-    private MoesifKeyRetriever keyRetriever;
     private MoesifClient client;
 
     public ParallelQueueWorker(BlockingQueue<MetricEventBuilder> queue, MoesifClient moesifClient) {
