+++ component/publisher-client/src/test/java/org/wso2/am/analytics/publisher/ErrorHandlingTestCase.java
@@ -37,6 +37,7 @@ import org.wso2.am.analytics.publisher.util.Constants;
 import org.wso2.am.analytics.publisher.util.TestUtils;
 import org.wso2.am.analytics.publisher.util.UnitTestAppender;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -55,7 +56,7 @@ public class ErrorHandlingTestCase {
         configMap.put(Constants.AUTH_API_TOKEN, "some_token");
         MetricReporter metricReporter = MetricReporterFactory.getInstance().createMetricReporter(configMap);
         CounterMetric metric = metricReporter.createCounterMetric("test-connection-counter", MetricSchema.RESPONSE);
-        List<String> messages = appender.getMessages();
+        List<String> messages = new ArrayList<>(appender.getMessages());
         Assert.assertTrue(TestUtils.isContains(messages, "Unrecoverable error occurred when creating Eventhub "
                                                          + "Client"), "Expected error hasn't logged in the "
                                   + "EventHubClientClass");
